fixed Double type in the Ball class

+++ src/main/java/oop/Ball.java
@@ -9,37 +9,37 @@ public class Ball {
 
   public Ball () {
     color = "white";
-    x = 0;
-    y = 0;
-    r = 10;
-    v = 0;
+    x = 0.0;
+    y = 0.0;
+    r = 10.0;
+    v = 0.0;
   }
 
-  public Ball (int x, int y) {
+  public Ball (Double x, Double y) {
     color = "white";
     this.x = x;
     this.y = y;
-    r = 10;
-    v = 0;
+    r = 10.0;
+    v = 0.0;
   }
 
-  public Ball (int x, int y, int r, String color) {
+  public Ball (Double x, Double y, Double r, String color) {
     this.color = color;
     this.x = x;
     this.y = y;
     this.r = r;
-    v = 0;
+    v = 0.0;
   }
 
   public void hit () {
-    v += 5;
+    v += 5.0;
   }
 
-  public void hit (int f) {
+  public void hit (Double f) {
     v += f;
   }
 
   public void stop () {
-    v = 0;
+    v = 0.0;
   }
 }
